Add more to existing NPCs and tests
PoisonDoctor and PosionNurse is now able to greet the player
and let the player know what they need to do to get a treatment.

Few more tests are added to the test suite. It will now check whether
Poison Doctor NPC and Poison Nurse NPC says what they should say.
In the Poison Healer For Player test, it will check the case where the
Player does not have enough money to get a treatment.

The player should not be able to get a treatment if they do not
have enough money, and the Poison Nurse NPC should clarify that
to the player.