Add HeavyStatusAttacker and fix tests
Added the HeavyStatusAttacker so attacking with the wand now uses the
heavy attacker class, rather than directly applying the slow. Changed
surrounding code to accommodate this, wand currently applied slow 100%
of the time.

Also changed format of the tests, HeavyStatusAttackerTest now only tests
the functionality of its corresponding class, rather than being
dependent on the functionality of the wand.

WandOfSluggishnessTest added a couple of new tests from
HeavyStatusAttacker, now more thoroughly tests the functionality of the
wand.

Overall minor changes in tests and working code to allow tests to pass.