Add new NPCs and tests
Two new doctor and nurse NPCs are added for the other two
states: confused and shocked. Along with the new NPCs, new quest
files are added as well. The doctors and nurses for different states
functions in a very similar way but the details such as their dialogues
are checked by the tests.

The deniran.xml file has been modified as well to make the new NPCs
appear in the game  screen.